220415 before css

+++ src/controllers/videoController.js
@@ -1,32 +1,53 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async(req,res) => {
     try{
-        const videos = await Video.find({}).sort({createdAt:"desc"});
-        return res.render("home",{pageTitle:"Home", videos});
+        const videos = await Video.find({})
+        .sort({ createdAt: "desc" })
+        .populate("owner");        return res.render("home",{pageTitle:"Home", videos});
     } catch(error){
         return res.status(400).render("server-error", error);
     }
 }
 export const watch = async(req, res) => {
     const {id}=req.params;
-    const video = await Video.findById(id);
+    //videoSchema의 owner 항목의 ref로 User를 넣어줬기 때문에
+    //populate를 이용해 owner를 id로 연결된 User로 채울 수 있음.
+    //object (여기서는 owner에 해당하는 user 정보) 전체를 불러와야 할 때 populate 사용]
+    const video = await Video.findById(id).populate("owner");
     return res.render("watch",{pageTitle:video.title, video});
 }
 export const getEdit = async(req, res) => {
-    const {id}=req.params;
+    const{
+        params:{id},
+        session:{
+            user:{_id},
+        }
+    }=req;
     const video = await Video.findById(id);
     if(!video){
         return res.status(404).render("404", {pageTitle: "Video not found."});
     }
+    if(String(video.owner)!==String(_id)){
+        return res.status(403).redirect("/");
+    }
     return res.render("edit",{pageTitle:`Edit: ${video.title}`, video});
 };
 export const postEdit = async(req, res) => {
-    const {id}=req.params;
-    const {title, description, hashtags}=req.body;
+    const {
+        params:{id},
+        body:{title, description, hashtags},
+        session:{
+            user:{_id},
+        }
+    }=req;
     const video = await Video.exists({ _id: id});
     if(!video){
         return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(String(video.owner)!==String(_id)){
+        return res.status(403).redirect("/");
     }
     await  Video.findByIdAndUpdate(id,{
         title,
@@ -40,13 +61,30 @@ export const getUpload = (req, res) => {
     return res.render("upload",{pageTitle:"Upload Video"});
 };
 export const postUpload = async(req, res) => {
-    const {title, description, hashtags}=req.body;
+    const {
+        session:{
+            user:{
+                _id,
+            },
+        },
+        file:{
+            path:fileUrl,
+        },
+        body:{
+            title, description, hashtags,
+        }
+    } = req;
     try{
-        await Video.create({
+        const newVideo = await Video.create({
+            owner:_id,
+            fileUrl,
             title,
             description,
             hashtags: Video.formatHashtags(hashtags),
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
     } catch(error){
         return res.status(400).render("upload",{
@@ -58,7 +96,19 @@ export const postUpload = async(req, res) => {
 };
 
 export const getDelete = async(req, res) => {
-    const {id}=req.params;
+    const {
+        params:{id},
+        session:{
+            user:{_id},
+        }
+    }=req;
+    const video=await findById(id);
+    if(!video){
+        return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(String(video.owner)!==String(_id)){
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 }
@@ -71,7 +121,7 @@ export const search = async(req, res) => {
             title: {
                 $regex: new RegExp(keyword,"i"),
             },
-        });
+        }).populate("owner");
     }
     return res.render("search",{pageTitle: "Search", videos});
 }
